backend/requirements.txt
# Django and related libraries
Django==5.1.2
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==3.13.0
django-axes==6.0.0

# Authentication and OAuth
djangorestframework-simplejwt==5.2.2
django-allauth==0.54.0
social-auth-app-django==5.0.0

# AI and OpenAI
openai==0.28.0  # Latest compatible version with your usage

# Spotify Integration and other API handling
spotipy==2.20.0
requests-oauthlib==1.3.1

# Data and AI Processing
scikit-learn==1.3.0
nltk==3.8.1
gensim==4.3.2

# API and Asynchronous Requests
requests==2.32.3
aiohttp==3.10.10
asyncio==3.4.3
httpx==0.27.2

# JSON Web Tokens and Security
cryptography==41.0.3
PyJWT==2.8.0

# Caching and Background Tasks
celery==5.2.7
redis>=5.2.0

# Web Scraping and Data Gathering
beautifulsoup4==4.12.2
scrapy==2.8.0
selenium==4.10.0
lxml==4.9.3

# Search Engine and NLP (Natural Language Processing)
whoosh==2.7.4
elasticsearch==8.10.1
langdetect==1.0.9

# Image Processing and Media
Pillow==10.0.0
moviepy==1.0.3

# Geolocation and Maps
geopy==2.3.0
folium==0.14.0

# Chatbot and Messenger Features
django-channels==0.7.0

# Data Validation and Parsers
pydantic==2.9.2
jsonschema==4.19.0

# Utility and Miscellaneous
python-dotenv==1.0.1
tqdm==4.66.6
bleach==6.0.0
